fix: Increase window for use of deencrypted key
Note: this is a port of https://github.com/oblador/react-native-keychain/pull/339.

This increases the time window you have to use the deencrypted key from 1 to 5 seconds
after the user unlocks their device when prompted. It seems that the time delay between
when face unlock on Android occurs and when the user chooses to continue back into the
app is enough that 1 second is too small of a window to work.

Taking a look, this is most likely due to how users choose to continue after facial
recognition on Android: the explicit confirmation step is probably what pushes them
over the time limit for use of the deencrypted key. This is further likely due to how
touch recognition works (you proceed immediately and automatically), though it can't
be confirmed until test devices are in our hands.