feat: Fallback to fingerprints only if face unlock fails
Some devices misreport their sensors as being strong enough to be used
with BiometricPrompt and the keystore, particularly Samsung.

These changes include a fallback that occurs on authentication error
while authenticating with face (checked based on presence of face feature
only due to lack of Android API features until Android 11). This fallback
forces the user to authenticate with fingerprints should they have any
enrolled on failure in the hopes of reducing the amount of dropoff
and confusion when using face.